Crate input_form

A Ratatui example that demonstrates how to handle input form focus

This example demonstrates how to handle cursor and input focus between multiple fields in a form. You can navigate between fields using the Tab key.

This does not handle cursor movement etc. This is just a simple example. In a real application, consider using tui-input, or tui-prompts, or tui-textarea.

This example runs with the Ratatui library code in the branch that you are currently reading. See the latest branch for the code which works with the most recent Ratatui release.

Structs

AgeField 🔒

A new-type representing a person’s age in years (0-130).

App 🔒

InputForm 🔒

StringField 🔒

A new-type representing a string field with a label.

Enums

AppState 🔒

Focus 🔒

Functions

main 🔒